requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2.4
django-cors-headers==2.5.3
djangorestframework==3.9.3
gunicorn==19.9.0
idna==2.8
lxml==4.3.3
pytz==2019.1
requests==2.21.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.24.3